feat(memories): move Memory classes into agents/utils/memories
