+++ rlberry/agents/avec/avec_ppo.py
@@ -5,26 +5,10 @@ import torch
 
 import gym.spaces as spaces
 from rlberry.agents import IncrementalAgent
-
+from rlberry.agents.utils.memories import Memory
 from rlberry.agents.utils.torch_models import ValueNet, PolicyNet
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
-
-class Memory:
-    def __init__(self):
-        self.actions = []
-        self.states = []
-        self.logprobs = []
-        self.rewards = []
-        self.is_terminals = []
-
-    def clear_memory(self):
-        del self.actions[:]
-        del self.states[:]
-        del self.logprobs[:]
-        del self.rewards[:]
-        del self.is_terminals[:]
 
 
 class AVECPPOAgent(IncrementalAgent):
