+++ rlberry/agents/cem/cem.py
@@ -7,34 +7,12 @@ import torch.nn as nn
 import rlberry.seeding as seeding
 import gym.spaces as spaces
 from rlberry.agents import Agent
+from rlberry.agents.utils.memories import CEMMemory
 
 # choose device
 from rlberry.agents.utils.torch_models import Net
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
-
-class Memory:
-    def __init__(self, max_size):
-        self.max_size = max_size
-        self.clear()
-
-    def clear(self):
-        self.size = 0
-        self.states = []
-        self.actions = []
-        self.rewards = []
-
-    def append(self, state, action, reward):
-        self.states.append(state)
-        self.actions.append(action)
-        self.rewards.append(reward)
-        self.size += 1
-        if self.size == self.max_size+1:
-            self.states.pop(0)
-            self.actions.pop(0)
-            self.rewards.pop(0)
-            self.size = self.max_size
 
 
 class CEMAgent(Agent):
@@ -102,7 +80,7 @@ class CEMAgent(Agent):
                                           lr=self.learning_rate)
 
         # memory
-        self.memory = Memory(self.batch_size)
+        self.memory = CEMMemory(self.batch_size)
 
         # logging config
         self._last_printed_ep = 0
